(modify) 인터페이스 및 구현 클래스 네이밍 변경
